updated marlin serialization
