+++ auto_gptq/nn_modules/qlinear/qlinear_marlin.py
@@ -166,19 +166,6 @@ class QuantLinear(nn.Module):
         C = C + self.bias if self.bias is not None else C 
         return C
 
-
-# Adapted from https://github.com/rib-2/marlin/tree/conversion
-def _validate_marlin_compatibility(quantization_config):
-    if quantization_config.bits != 4:
-        return f"The quantized model uses a bitwidth different than 4 (found {quantization_config.bits})"
-    if quantization_config.group_size != 128 and quantization_config.group_size != -1:
-        return f"The quantized model uses a group size that is not 128 or -1 (found quantization_config.group_size)"
-    if not quantization_config.sym:
-        return f"The quantized model uses asymmetric quantization"
-    if quantization_config.desc_act:
-        return f"The quantized model uses act-order (also called desc-act) scheme"
-    return None
-
 # Copied from https://github.com/IST-DASLab/marlin/pull/1
 @torch.no_grad()
 def unpack_4bit_to_32bit_signed(qweight, qzeros):
